app_fastapi/requirements.txt
mlflow==2.21.3
scikit-learn==1.4.2
transformers==4.38.0
torch==2.2.0
tokenizers>=0.15.2
requests>=2.31.0,<3
numpy==1.24.4
fastapi
uvicorn[standard]
pydantic
pandas
gunicorn
openpyxl
boto3
python-multipart
python-dotenv
xgboost
spacy